Add assets as global state
