+++ src/components/portfoliosection/PortfolioComponent.js
@@ -1,47 +1,9 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
 import Col from "react-bootstrap/Col";
 import Table from "react-bootstrap/Table";
 import "./portfolio.css";
 import Asset from "./asset/Asset";
-import { assetData } from "../../api";
 
-function PortfolioComponent() {
-  const [assets, setAssets] = useState([]);
-
-  useEffect(() => {
-    async function getAssets() {
-      const response = await axios.get("http://localhost:8080/api/assets");
-      const requestedAssets = response.data;
-      console.log(requestedAssets);
-
-      Promise.all(
-        requestedAssets.map(async function (asset) {
-          try {
-            const response = await axios.get(assetData(asset.AssetId));
-            let cp = response.data.market_data.current_price.eur;
-            let value = Number(cp) * Number(asset.Amount);
-            return { ...asset, value: value, price: cp };
-          } catch (error) {
-            console.log(error.response.data.error);
-            throw error;
-          }
-        })
-      )
-        .then((newAssetArray) => {
-          setAssets(newAssetArray);
-          console.log(newAssetArray);
-          console.log(assets);
-        })
-        .catch((error) => {
-          console.log(error);
-        });
-    }
-    let interval = setTimeout(() => {
-      getAssets();
-    }, 3000);
-  }, [assets]);
-
+function PortfolioComponent({ assets }) {
   return (
     <Col className="section mt-5">
       <h1>Portfolio</h1>
